+++ src/jira-api.js
@@ -381,7 +381,7 @@ export class JiraClient {
       'customfield_10016',
       'customfield_10026',
       'subtasks', 'created', 'updated',
-      'duedate', 'startDate', 'rank',
+      'duedate', 'startDate', 'customfield_10015', 'rank',
       'labels'
     ];
     if (withWorklogs) fields.push('worklog');